backend/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.5

SQLAlchemy==2.0.23
python-dotenv==1.0.0
PyJWT==2.8.0

openai==1.3.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3

bcrypt==4.1.1
Werkzeug==3.0.1

gunicorn==21.2.0
psycopg2-binary==2.9.9



